requirements.txt
Chroma==0.2.0
chroma-hnswlib==0.7.3
chromadb==0.5.3
gradio==4.43.0
ipykernel==6.29.5
langchain==0.2.16
langchain-chroma==0.1.3
langchain-community==0.2.16
langchain-core==0.2.38
langchain-openai==0.1.23
langchain-text-splitters==0.2.4
langgraph==0.2.19
langgraph-checkpoint==1.0.9
langsmith==0.1.116
numpy==1.26.4
openai==1.44.0
pandas==2.2.2
pydantic==2.9.0
pydantic_core==2.23.2
pypdf==4.3.1
pyprojroot==0.3.0
PyYAML==6.0.2
SQLAlchemy==2.0.34



